feat: add module permissions middleware
